Add user node-red to dialout group
This is needed for the Serialport-Node to work properly.
At the moment I do it manually after each update. But I think it would be very convinient also for other users to have this baked in.
See this Issue: https://github.com/node-red/node-red-docker/issues/15

Other option could be to use a environment variable for this? 
e.g. which groups the user node-red should belong to.